requirements.txt
bidict==0.21.4
cffi==1.15.0
click==8.0.3
clipspy==1.0.0
colorama==0.4.4
Flask==2.0.2
Flask-SocketIO==5.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
python-engineio==4.3.0
python-socketio==5.5.0
Werkzeug==2.0.2
